[materialsystem] Fixed ShaderAPITextureHandle_t, VertexShader_t, PixelShader_t definitions for 64 bits [stdshaders] Fixed commandbuilder for 64bits [tier1] CUtlFixedLinkedList now uses intp instead of int

+++ src/public/shaderapi/ishaderdynamic.h
@@ -18,7 +18,7 @@
 #include "materialsystem/imaterialsystem.h"
 
 
-typedef int ShaderAPITextureHandle_t;
+typedef intp ShaderAPITextureHandle_t;
 
 //-----------------------------------------------------------------------------
 // forward declarations
